inflearn/inflearn01/src/main/java/com/inflearn01/inflearn01/service/MemberService.java
package com.inflearn01.inflearn01.service;

import com.inflearn01.inflearn01.domain.Member;
import com.inflearn01.inflearn01.repository.MemberRepository;
import com.inflearn01.inflearn01.repository.MemoryMemberRepository;

import java.util.List;
import java.util.Optional;

public class MemberService {
    // 서비스클래스는 비즈니스에 가까운 용어를 써야한다.
    private final MemberRepository memberRepository;
    public MemberService(MemberRepository memberRepository){
        this.memberRepository = memberRepository;
    }
    /*회원가입*/

    public Long join(Member member){
        validateDuplicateMember(member); // 중복 회원 검증
        memberRepository.save(member);
        return member.getId();
    }

    private void validateDuplicateMember(Member member) {
        memberRepository.findByName(member.getName())
            .ifPresent(m -> {
            throw new IllegalStateException("이미 존재하는 회원입니다.");
        });
    }

    public List<Member> findMembers(){
        return memberRepository.findAll();
    }

    public Optional<Member> findOne(Long memberId){
        return memberRepository.findById(memberId);
    }
}
